bump electrum_client version in bdk_electrum
This bumps the electrum_client version used by bdk_electrum and
adds corresponding features that control the provider used by
rustls.

These changes are based on the version of this file in BDK's
v1.0.0-beta.2 tag.

The motivation is to be able to avoid using the default
aws-lc-rs provider from rustls when using electrum_client via the
bdk_electrum crate.